requirements.txt
PyInstaller
pandas==1.2.4
PyQt5==5.15.4
PyQt5-stubs==5.15.2.0
openpyxl==3.0.7
python-dateutil
lxml==4.6.3